split-utils * Move BREAKPOINTS to constants.ts * Move EMAIL_REGEX to AddNewOwnersInput.tsx
