+++ src/dashboards/BusStop/BusStopDashboard.tsx
@@ -4,7 +4,7 @@ import { useLocation, useParams } from 'react-router-dom'
 import { useLongPress } from 'use-long-press'
 import { Loader } from '@entur/loader'
 import { FormFactor } from '@entur/sdk/lib/mobility/types'
-import { BREAKPOINTS } from '../../utils'
+import { BREAKPOINTS } from '../../constants'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
 import { DEFAULT_ZOOM } from '../../constants'
 import {
