+++ src/dashboards/Timeline/TimelineDashboard.tsx
@@ -6,7 +6,7 @@ import { LegBone } from '@entur/travel'
 import { LegMode } from '@entur/sdk'
 import { colors } from '@entur/tokens'
 import { WalkingIcon } from '@entur/icons'
-import { BREAKPOINTS } from '../../utils'
+import { BREAKPOINTS } from '../../constants'
 import { LineData, IconColorType } from '../../types'
 import { useStopPlacesWithDepartures, useWalkInfo } from '../../logic'
 import { DashboardWrapper } from '../../containers/DashboardWrapper/DashboardWrapper'
